Move handle_merge_reply into autocompound.rs
